Replace Yomichan with Yomitan (the latest fork)
This commit changes the setup from Yomichan to Yomitan. Wherever the text or the pictures referred to Yomichan directly, I either kept the Yomichan reference or I changed it to Yomitan/Yomichan. This notably fixes the fact you cannot download Yomichan from FooSoft directly anymore (and the myriad of bugs with it), and it makes sure the basic setup is correct. I also changed the reference to Kanji Grid because the version linked doesn't work anymore.